Move utxoForAddresses v2 to utxoAtPoint
